+++ resources/views/device-overview-panel.blade.php
@@ -1,0 +1,3 @@
+<x-device.overview.panel title="NETCONF" icon="fa fa-terminal" :href="$href">
+    @include('netconf::device-overview-body')
+</x-device.overview.panel>